% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apra_stat_pubs}
\alias{apra_stat_pubs}
\title{Statistical Publications Produced by APRA}
\format{
A data frame with 7 rows and 3 variables:
\describe{
\item{stat_pub_name}{the full name of the statistical publication.}
\item{stat_pub_acronym}{the acronym of the statistical publication.}
\item{cur_hist}{available current and historical versions of the statistical publication.}
}
}
\usage{
apra_stat_pubs
}
\description{
This dataset contains a record of all statistical publications produced by
APRA that are available in \code{readapra}.
}
\keyword{datasets}
