% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apra_statistical_publications}
\alias{apra_statistical_publications}
\title{Statistical Publications Produced by APRA}
\format{
A data frame with 7 rows and 3 variables:
\describe{
\item{statistical_publication_name}{the full names of the statistical publications}
\item{statistical_publication_acronym}{the acronyms of the statistical publications}
\item{cur_hist}{available current and historical versions the statistical publications}
}
}
\usage{
apra_statistical_publications
}
\description{
This dataset contains a list of all statistical publications produced by
APRA that are available in \code{readapra}.
}
\keyword{datasets}
