% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_madis.R
\name{clean_madis_raw}
\alias{clean_madis_raw}
\title{This function imports the raw MADIS data as a tibble and then proceeds to
tidy it into a usable format for the end user.}
\usage{
clean_madis_raw(file_path, cur_hist, call = rlang::caller_env())
}
\description{
This function imports the raw MADIS data as a tibble and then proceeds to
tidy it into a usable format for the end user.
}
\keyword{internal}
