% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadips.R
\name{qadip_key_stats_names}
\alias{qadip_key_stats_names}
\title{Gets dependency names for the Key Stats sheet in the QADIP. This sheet is
structured differently from the other sheets and needs a lot of tweaking to
get working.}
\usage{
qadip_key_stats_names(tidyxl_data, series_hierarchy_data)
}
\arguments{
\item{tidyxl_data}{The QADIP data sourced using the tidyxl package}

\item{series_dependency_data}{The series dependency data sourced using
get_series_dependencies()}
}
\description{
Gets dependency names for the Key Stats sheet in the QADIP. This sheet is
structured differently from the other sheets and needs a lot of tweaking to
get working.
}
\keyword{internal}
