% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adipops.R
\name{read_adipops}
\alias{read_adipops}
\title{Read ADI Points of Presence Statistics}
\usage{
read_adipops()
}
\value{
A tibble containing the ADIPOPS data
}
\description{
Read in the Authorised Deposit-taking Institution Points of Presence
Statistics (ADIPOPS) from APRA's website.
}
\examples{
read_adipops("current")
}
