% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adipops.R
\name{read_adipops_local}
\alias{read_adipops_local}
\title{Read ADI Points of Presence Statistics locally}
\usage{
read_adipops_local(file_path)
}
\arguments{
\item{file_path}{path to the local .xlsx file.}
}
\value{
A tibble containing the ADI Points of Presence Statistics data.
}
\description{
Import the Authorised Deposit-taking Institution Points of Presence
Read ADI Points of Presence Statistics from a local file.
}
\examples{
\donttest{
# Downloading the latest ADIPOPS file
adipops_file_path <- download_apra(publication = "adipops")

# Importing the data into R
read_adipops_local(file_path = adipops_file_path)
}
}
