% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_madis.R
\name{read_madis_local}
\alias{read_madis_local}
\title{Read Monthly ADI Statistics locally}
\usage{
read_madis_local(file_path, cur_hist)
}
\arguments{
\item{file_path}{path to the local .xlsx file.}

\item{cur_hist}{character vector determining whether to download the current
publication (\code{"current"}) or the historic publication (\code{"historic"}).}
}
\value{
A tibble containing the Monthly ADI Statistics data.
}
\description{
Import the Monthly Authorised Deposit-taking Institution Statistics (MADIS)
from a local file. Both the current and historic versions of this
statistical publication are available.
}
\examples{
\donttest{

# Downloading the latest MADIS file
madis_file_path <- download_apra(publication = "madis")

# Importing the data into R
read_madis_local(file_path = madis_file_path, cur_hist = "current")
}
}
