% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadicp.R
\name{read_qadicp}
\alias{read_qadicp}
\title{Read Quarterly ADI Centralised Publication}
\usage{
read_qadicp(path = tempdir(), overwrite = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{path}{path to where the downloaded file should be saved. Uses
\code{\link[base:tempfile]{base::tempdir()}} by default.}

\item{overwrite}{whether to overwrite the downloaded file when re-downloading
the file.}

\item{quiet}{whether to suppress the download progress bar.}

\item{...}{additional arguments to be passed to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
A tibble containing the Quarterly ADI Centralised Publication data.
}
\description{
Download and import the Quarterly Authorised Deposit-taking
Institution Centralised Publication (QADICP) from APRA's website.
}
\examples{
\dontrun{
read_qadicp()
}
}
