% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadicp.R
\name{read_qadicp_local}
\alias{read_qadicp_local}
\title{Read Quarterly ADI Centralised Publication locally}
\usage{
read_qadicp_local(file_path)
}
\arguments{
\item{file_path}{path to the .xlsx file.}
}
\value{
A tibble containing the Quarterly ADI Centralised Publication data.
}
\description{
Import the Quarterly Authorised Deposit-taking Institution Centralised
Publication (QADICP) from a local file.
}
\examples{
\dontrun{
read_qadicp_local(file_path = ~ path / to / xlsx / file)
}
}
