% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadicp.R
\name{read_qadicps_local}
\alias{read_qadicps_local}
\title{Read Quarterly ADI Centralised Publication locally}
\usage{
read_qadicps_local(file_path)
}
\arguments{
\item{file_path}{path to the local .xlsx file.}
}
\value{
A tibble containing the Quarterly ADI Centralised Publication data.
}
\description{
Import the Quarterly Authorised Deposit-taking Institution Centralised
Publication (QADICP) from a local file.
}
\examples{
\donttest{
# Downloading the latest QADICP file
qadicp_file_path <- download_apra(publication = "qadicp")

# Importing the data into R
read_qadicp_local(file_path = qadicp_file_path)
}
}
