% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadipexs.R
\name{read_qadipexs}
\alias{read_qadipexs}
\title{Read Quarterly ADI Property Exposure Statistics}
\usage{
read_qadipexs(cur_hist, path = tempdir(), overwrite = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{cur_hist}{character vector determining whether to download the current
publication (\code{"current"}) or the historic publication (\code{"historic"}).}

\item{path}{path to where the downloaded file should be saved. Uses
\code{\link[base:tempfile]{base::tempdir()}} by default.}

\item{overwrite}{whether to overwrite the downloaded file when re-downloading
the file.}

\item{quiet}{whether to suppress the download progress bar.}

\item{...}{additional arguments to be passed to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
A tibble containing the Quarterly ADI Property Exposure Statistics
data.
}
\description{
Download and import the Quarterly Authorised Deposit-taking Institution
Property Exposure Statistics (QADIPEXS) from APRA's website. Both the
current and historic versions of this statistical publication are available.
}
\examples{
\dontrun{
read_qadipexs(cur_hist = "current")
}
}
