% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadips.R
\name{read_qadips}
\alias{read_qadips}
\title{Read Quarterly ADI Performance Statistics}
\usage{
read_qadips()
}
\value{
A tibble containing the Quarterly ADI Performance Statistics data.
}
\description{
Download and import the Quarterly Authorised Deposit-taking Institution
Performance Statistics (QADIPS) from APRA's website.
}
\examples{
read_qadips()
}
