% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadips.R
\name{read_qadips_local}
\alias{read_qadips_local}
\title{Read Quarterly ADI Performance Statistics locally}
\usage{
read_qadips_local(file_path)
}
\arguments{
\item{file_path}{path to the local .xlsx file.}
}
\value{
A tibble containing the Quarterly ADI Performance Statistics data.
}
\description{
Import the Quarterly Authorised Deposit-taking Institution
Performance Statistics (QADIPS) from a local file.
}
\examples{
\donttest{
# Downloading the latest QADIPS file
qadips_file_path <- download_apra(publication = "qadips")

# Importing the data into R
read_qadips_local(file_path = qadips_file_path)
}
}
