% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpexs.R
\name{read_qpexs}
\alias{read_qpexs}
\title{Read Quarterly ADI Property Exposure Statistics}
\usage{
read_qpexs(cur_hist)
}
\arguments{
\item{cur_hist}{Whether to access the current or historical series. Valid
values are `"current"` and `"historical"`.}
}
\value{
A tibble containing the Quarterly ADI Property Exposure Statistics
data.
}
\description{
Download and import the Quarterly Authorised Deposit-taking Institution
Property Exposure Statistics (QPEXS) from APRA's website.
}
\examples{
read_qpexs("current")
}
