% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_apra.R
\name{download_apra}
\alias{download_apra}
\title{Download a Statistical Publication File from APRA's Website}
\usage{
download_apra(
  publication,
  cur_hist = "current",
  path = tempdir(),
  overwrite = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{publication}{character vector containing the acronym of the statistical
publication you want to download.}

\item{cur_hist}{character vector determining whether to download the current
publication (\code{"current"}) or the historic publication (\code{"historic"}). Please
note that not all statistical publications have a historic publication. By
default this is set to (\code{"current"}).}

\item{path}{path to where the downloaded file should be saved. Uses
\code{\link[base:tempfile]{base::tempdir()}} by default.}

\item{overwrite}{whether to overwrite the downloaded file when re-downloading
the file.}

\item{quiet}{whether to suppress the download progress bar.}

\item{...}{additional arguments to be passed to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
A character vector detailing the file path to the downloaded file.
}
\description{
Downloads a specified statistical publication file from APRA's website. By
default files are saved to a temporary directory.
}
\examples{
\donttest{
download_path <- download_apra(publication = "qadips", cur_hist = "current")

print(download_path)
}
}
