% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adipops.R
\name{read_adipops}
\alias{read_adipops}
\title{Read ADI Points of Presence Statistics}
\usage{
read_adipops(path = tempdir(), overwrite = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{path}{path to where the downloaded file should be saved. Uses
\code{\link[base:tempfile]{base::tempdir()}} by default.}

\item{overwrite}{whether to overwrite the downloaded file when re-downloading
the file.}

\item{quiet}{whether to suppress the download progress bar.}

\item{...}{additional arguments to be passed to \code{\link[utils:download.file]{utils::download.file()}}.}

\item{cur_hist}{character vector determining whether to download the current
publication (\code{"current"}) or the historic publication (\code{"historic"}).}
}
\value{
A tibble containing the ADIPOPS data.
}
\description{
Download and import the Authorised Deposit-taking Institution Points of
Presence Statistics (ADIPOPS) from APRA's website.
}
\examples{
\dontrun{
read_adipops("current")
}
}
