% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adipops.R
\name{read_adipops_local}
\alias{read_adipops_local}
\title{Read ADI Points of Presence Statistics locally}
\usage{
read_adipops_local()
}
\value{
A tibble containing the ADIPOPS data.
}
\description{
Read in the Authorised Deposit-taking Institution Points of Presence
Statistics (ADIPOPS) from a local file.
}
\examples{
\dontrun{
read_adipops_local(file_path = ~path/to/xlsx/file)
}
}
