% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_madis.R
\name{read_madis}
\alias{read_madis}
\title{Read Monthly ADI Statistics}
\usage{
read_madis(cur_hist)
}
\arguments{
\item{cur_hist}{Character; valid values are `"current"` or `"historical"`.}
}
\value{
A tibble containing the Monthly ADI Statistics data
}
\description{
Read in the Monthly Authorised Deposit-taking Institution Statistics (MADIS)
from APRA's website.
}
