% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_madis.R
\name{read_madis_local}
\alias{read_madis_local}
\title{Read Monthly ADI Statistics locally}
\usage{
read_madis_local(file_path, cur_hist)
}
\arguments{
\item{file_path}{The file path to the local MADIS .xlsx file.}

\item{cur_hist}{Whether to access the current or historical series. Valid
values are `"current"` and `"historical"`.}
}
\value{
A tibble containing the Monthly ADI Statistics data.
}
\description{
Read in the Monthly Authorised Deposit-taking Institution Statistics (MADIS)
from a local file.
}
\examples{
\dontrun{
read_madis_local(file_path = ~path/to/xlsx/file, cur_hist = "current")
}
}
