% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadicp.R
\name{read_qadicp}
\alias{read_qadicp}
\title{Read Quarterly ADI Centralised Publication}
\usage{
read_qadicp()
}
\value{
A tibble containing the Quarterly ADI Centralised Publication data.
}
\description{
Download and import Read in the Quarterly Authorised Deposit-taking
Institution Centralised Publication (QADICP) from APRA's website.
}
\examples{
read_qadicp()
}
