% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadipexs.R
\name{read_qadipexs_local}
\alias{read_qadipexs_local}
\title{Read Quarterly ADI Property Exposure Statistics locally}
\usage{
read_qadipexs_local(file_path, cur_hist)
}
\arguments{
\item{file_path}{path to the local .xlsx file.}

\item{cur_hist}{character vector determining whether to download the current
publication (\code{"current"}) or the historic publication (\code{"historic"}).}
}
\value{
A tibble containing the Quarterly ADI Performance Statistics data.
}
\description{
Import the Quarterly Authorised Deposit-taking Institution
Property Exposure Statistics (QPEXS) from a local file. Both the current and
historic versions of this statistical publication are available.
}
\examples{
\donttest{
# Downloading the latest QADIPEXS file
qadipexs_file_path <- download_apra(publication = "qadipexs")

# Importing the data into R
read_qadipexs_local(file_path = qadipexs_file_path, cur_hist = "current")
}
}
