% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qadipexs.R
\name{read_qadipexs_local}
\alias{read_qadipexs_local}
\title{Read Quarterly ADI Property Exposure Statistics locally}
\usage{
read_qadipexs_local(file_path)
}
\arguments{
\item{file_path}{The file path to the local QPEXS .xlsx file.}
}
\value{
A tibble containing the Quarterly ADI Performance Statistics data.
}
\description{
Import the Quarterly Authorised Deposit-taking Institution
Property Exposure Statistics (QPEXS) from a local file.
}
\examples{
\dontrun{
read_qpexs_local(file_path = ~path/to/xlsx/file)
}
}
