% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readapra_package.R
\docType{package}
\name{readapra-package}
\alias{readapra}
\alias{readapra-package}
\title{readapra: Download and Tidy Data from the Australian Prudential Regulation Authority}
\description{
Download the latest data from the Australian Prudential Regulation Authority \url{https://www.apra.gov.au/} and import it into R as a tidy data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/javanderwal/readapra}
  \item Report bugs at \url{https://github.com/javanderwal/readapra/issues}
}

}
\author{
\strong{Maintainer}: Jarrod van der Wal \email{jarvanderwal@gmail.com} [copyright holder]

}
\keyword{internal}
